% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latest_gefs.R
\name{get_latest_gefs}
\alias{get_latest_gefs}
\title{Mengunduh GEFS Terkini}
\usage{

  get_latest_gefs(var = "Total_precipitation_surface_6_Hour_Accumulation_ens",
  query = "&north=6&west=90&east=141&south=-11&temporal=all",
  destfile = "latest_gefs.nc")
}
\arguments{
\item{var}{nama variabel, jika lebih dari satu pisahkan dengan tanda koma tanpa spasi}

\item{query}{tambahan query}

\item{destfile}{destinasi berkas GEFS unduhan}
}
\value{
berkas GEFS dan URL NCSS
}
\description{
Fungsi untuk mengunduh GEFS terkini dengan memanfaatkan NetCDF Subset Service
}
