% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latest_gfs.R
\name{get_latest_gfs}
\alias{get_latest_gfs}
\title{Mengunduh GFS 0.25 Terkini}
\usage{

  get_latest_gfs(var = "Total_precipitation_surface_Mixed_intervals_Accumulation",
  query = "&north=6&west=90&east=141&south=-11&temporal=all",
  destfile = "latest_gfs.nc")
}
\arguments{
\item{var}{nama variabel, jika lebih dari satu pisahkan dengan tanda koma tanpa spasi}

\item{query}{tambahan query}

\item{destfile}{destinasi berkas GFS unduhan}
}
\value{
berkas GFS dan URL NCSS
}
\description{
Fungsi untuk mengunduh GFS 0.25 terkini dengan memanfaatkan NetCDF Subset Service
}
