% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_ncss.R
\name{query_ncss}
\alias{query_ncss}
\title{Antarmuka The NetCDF Subset Service (NCSS)}
\usage{
query_ncss(host = "thredds.ucar.edu", context = "thredds",
  service = "ncss",
  dataset = "grib/NCEP/GEFS/Global_1p0deg_Ensemble/members/GEFS_Global_1p0deg_Ensemble_20190213_0000.grib2",
  var = "Total_precipitation_surface_6_Hour_Accumulation_ens",
  query = "&north=6&west=90&east=141&south=-11&temporal=all")
}
\arguments{
\item{host}{nama peladen, misalnya thredds.ucar.edu}

\item{context}{"thredds"}

\item{service}{"ncss"}

\item{dataset}{path dataset dari katalog}

\item{var}{nama variabel}

\item{query}{tambahan query}
}
\value{
kueri NCSS
}
\description{
Fungsi antarmuka dari NetCDF Subset Service
}
