% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nc.R
\name{read_nc}
\alias{read_nc}
\title{Membaca netCDF gfs.0p25}
\usage{
read_nc(path)
}
\arguments{
\item{path}{lokasi dokumen gfs.0p25 format nc}
}
\value{
dataframe
}
\description{
Menghasilkan dataframe untuk parameter akumulasi presipitasi, temperatur minimum, temperatur maksimum, lintang, dan bujur yang berasal dari data gfs.0p25.
}
